Support for snippets
There is an snippets/ directory where we can drop any source file. In the
slide source any <snippet> tag will be replaced with the content of that file.

For example, with

  .slide
    %h1 Some title

    %snippet example.rb

The slides generator will load snippets/example.rb and include it where the
<snippet> tag was found. If pygmentize is installed in the system, it will be
formatted. If pygmentize can not be used the code is inserted in a <pre>.