Add <asset> tag
The slide can use this pseudo-tag to create links to the assets.

If the asset is an image, a <img> will be created.

If the asset is a javascript file, a <script> tag will be created.

If the asset is a CSS file, a <link> tag will be created.

If the type can not be determined, a simple link will be created.

The type can be forced adding a .style, .javascript or .image tag. Something
like

  %asset.image foo.apng