Use Shower instead of deck.js to control the slides.
Shower is smaller and easier to adapt, so it fits better in HighCarb.

The Rack application now provides a /vendor endpoint to provide the files for
the presentation management.

The styles are very basic. It is expected that the presentation defines its own
theme.

The /slides endpoint is also removed. Now, the generated HTML is embedded in the
response.