Custom Haml filters support.
In the `config.yaml` file of each presentation it is possible to register
programs that can be used to generate HTML from text in the Haml sources.

This feature is not officially supported by Haml, so may be broken in a future
version of the library.

The implementation of `HighCarb::HamlFilters` was adapted from the `:plain`
filter.