Stream protocol for custom filters.
If a custom filter is configured to use this protocol, the program will run as a
background process, and messages between HighCarb and it are sent using a simple
protocol (bytesize + content).

This avoids big performance issues when a presentation contains a lot of filter
instances, and the program to render them is slow to load.